First batch of changes regarding Solver logs management and pathing
