+++ sparkle/tools/__init__.py
@@ -1,6 +1,6 @@
 """Init for the tools module."""
 from sparkle.tools.general import get_time_pid_random_string
 from sparkle.tools.pcsparser import PCSParser
-from sparkle.tools.runsolver_parsing import get_solver_args
 from sparkle.tools.slurm_parsing import SlurmBatch
 from sparkle.tools.solver_wrapper_parsing import get_solver_call_params
+from sparkle.tools.runsolver import RunSolver
